refactor(patrol_cli): extract abstract VideoRecordingManager base class
- Shared logic (operation serialization queue, handleTestEntry, dispose,
  test name sanitization) moves to an abstract VideoRecordingManager in
  crossplatform/; Android and iOS managers now only implement platform
  start/stop
- Replace the inline onLogEntry closures in both test backends with
  VideoRecordingManager.wrapOnLogEntry
- Drop redundant comments around video config construction
