+++ packages/patrol_cli/lib/src/commands/test.dart
@@ -490,7 +490,6 @@ See https://github.com/leancodepl/patrol/issues/1316 to learn more.
     Future<void> Function() action;
     Future<void> Function()? finalizer;
 
-    // Video output directory always follows test directory
     final videoConfig = VideoRecordingConfig(
       enabled: boolArg('record-video'),
       outputDirectory: stringArg('video-output-dir') ?? '$testDirectory/videos',
